# Make casting of string to integral throw if input contains space