This change makes it throw exception to cast string to integral
types (bigint, integer, boolean, etc), when cast_to_int_by_truncate=false,
to have the same behavior as Presto. For the CAST expression, previous
behavior is to accept leading and trailing whitespaces in input strings
and can cast successfully when cast_to_int_by_truncate=false. 

For Presto, casting string to real & double accepts leading and trailing
whitespaces. So should keep the behavior as it is.

CAST expression accepts whitespace in input because it utilizes folly to do
conversion. folly explicitly accepts and processes leading and trailing
whitespaces of input strings. So we just need to override this part of behavior
and continue to utilize the rest.
https://github.com/facebook/folly/blob/00b7e630bdbd09e7f1cd173c80c2f52b1e631df6/folly/Conv.h#L1501-L1510
https://github.com/facebook/folly/blob/00b7e630bdbd09e7f1cd173c80c2f52b1e631df6/folly/Conv.h#L1458-L1466

Also add benchmark for the invalid cases of try_cast() and try(cast)) on strings
with whitespace.