@laithsakka 2023-10-24
we chatted offline