@mbasmanova 2023-10-24
> we chatted offline

@laithsakka @gggrace14 Laith, Ge, would you summarize this discussion?