@gggrace14 2023-10-25
> > we chatted offline
> 
> @laithsakka @gggrace14 Laith, Ge, would you summarize this discussion?

This change returns exception for casting strings with whitespace, which previously didn't throw. Talking with Laith for perf and reading his blog, I understand
1. To return exception, I can directly set the result to null or set the result to exception_ptr in `CastExpr::applyCastKernel()`, doing the same as we handle the empty string now. This avoids calling VELOX_CHECK, constructing Velox exception or throwing & rethrowing exceptions from folly, which are expensive.
2. For benchmarking, try_cast invalid case should be ~1.x ms. cast valid case should be kept ~2.x ms.