@gggrace14 2023-10-25
For benchmarking, after the change we should expect
- The valid cases `try_cast_valid & tryexpr_cast_valid & cast_valid` doesn't see much increase to time.
- The whitespace cases `try_cast_invalid_whitespace & tryexpr_cast_invalid_whitespace` has comparable time as the empty input cases `try_cast_invalid_empty_input & tryexpr_cast_invalid_empty_input`.

Before the change
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          1.47ms    678.21
cast##tryexpr_cast_invalid_empty_input                    326.49ms      3.06
cast##try_cast_valid                                        2.17ms    461.16
cast##tryexpr_cast_valid                                    2.14ms    468.35
cast##cast_valid                                            2.13ms    470.30

After the change
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##try_cast_invalid_empty_input                          1.35ms    740.15
cast##tryexpr_cast_invalid_empty_input                    344.44ms      2.90
cast##try_cast_invalid_whitespace                           1.45ms    689.42
cast##tryexpr_cast_invalid_whitespace                     335.06ms      2.98
cast##try_cast_valid                                        2.10ms    475.13
cast##tryexpr_cast_valid                                    2.19ms    456.77
cast##cast_valid                                            2.18ms    459.71