@gggrace14 2023-11-08
After investigation and discussion, decide not to proceed with this change, as ANSI SQL standard accepts leading and trailing whitespaces which is current behavior. Follow-up, https://github.com/prestodb/presto/issues/21315